backend/requirements.txt
flask
flask-cors
bcrypt
pycryptodome
requests
psutil
python-dotenv
gunicorn
firebase-admin
